test(workspace): cover get_workspace slug lookup
